+++ lib/Update.ps1
@@ -167,9 +167,14 @@ function Update-Scoop {
     .SYNOPSIS
         Update scoop itself and all buckets.
     #>
-    param()
+    param([Switch] $CheckLastUpdate)
 
     if (!(Test-CommandAvailable -Command 'git')) { Stop-ScoopExecution -Message 'Scoop uses Git to update itself. Run ''scoop install git'' and try again.' }
+    # Skip updates if not needed
+    if ($CheckLastUpdate -and ($false -eq (is_scoop_outdated))) {
+        return
+    }
+
     Write-UserMessage -Message 'Updating Scoop...' -Output
 
     # TODO: CONFIG refactor adoption