+++ libexec/scoop-uninstall.ps1
@@ -38,7 +38,7 @@ $Purge = $Options.p -or $Options.purge
 if (!$Applications) { Stop-ScoopExecution -Message 'Parameter <APP> missing' -Usage (my_usage) }
 if ($Global -and !(is_admin)) { Stop-ScoopExecution -Message 'Administrator privileges are required to uninstall globally installed applications.' -ExitCode 4 }
 
-if ($Applications -eq 'scoop') {
+if ($Applications -contains 'scoop') {
     & (Join-Path $PSScriptRoot '..\bin\uninstall.ps1') $Global $Purge
     exit $LASTEXITCODE
 }