+++ supporting/completion/scoop.lua
@@ -239,7 +239,7 @@ local scoopParser = parser({
     'virustotal' .. parser({getLocallyAvailableApplicationsByScoop},
         '-a' .. architectureParser, '--arch' .. architectureParser,
         '-s', '--scan',
-        '-n', '--no-depends',
+        '-i', '--independent',
         '-h', '--help'
     ):loop(1),
     'which' .. parser({'-h', '--help'})